Add console type
The console type is representing all kind of operation related to console.
and then we smash these data types into Free[F,A]
That is Free[Console, A] is defined.
And then call runFree that takes any F, converts it to target G which is a monad and run it